Move group page to (overview) and add loading page
